requirements.txt
Flask==2.2.2
Flask_Cors==5.0.0
gunicorn==23.0.0
numpy==1.24.3
opencv_python==4.10.0.84
Pillow==10.4.0
PyAutoGUI==0.9.54
python-dotenv==1.0.1
Requests==2.32.3
torch==2.3.1
Werkzeug==2.2.2
ultralytics==8.2.87
